Test the components
