+++ src/__Tests__/storeMock.js
@@ -1,0 +1,15 @@
+import { createStore, combineReducers, applyMiddleware } from 'redux';
+import thunk from 'redux-thunk';
+import logger from 'redux-logger';
+import statsReducer from './reduxMock';
+
+const reducer = combineReducers({
+  statsReducer,
+});
+
+const store = createStore(
+  reducer,
+  applyMiddleware(logger, thunk),
+);
+
+export default store;
